# Faster inverse of 4th order Tensors