--- a/src/math_ops.jl
+++ b/src/math_ops.jl
@@ -150,11 +150,11 @@ end
 end
 
 function Base.inv(t::Tensor{4, dim}) where {dim}
-    fromvoigt(Tensor{4, dim}, inv(tovoigt(t)))
+    fromvoigt(Tensor{4, dim}, inv(tovoigt(SMatrix, t)))
 end
 
 function Base.inv(t::SymmetricTensor{4, dim, T}) where {dim, T}
-    frommandel(SymmetricTensor{4, dim}, inv(tomandel(t)))
+    frommandel(SymmetricTensor{4, dim}, inv(tomandel(SMatrix, t)))
 end
 
 Base.:\(S1::SecondOrderTensor, S2::AbstractTensor) = inv(S1) ⋅ S2