Profit from fast + allocation free `StaticArrays` inverses
```julia
julia> A = rand(Tensor{4,3}); Asym = symmetric(A);       

# on master
julia> @btime inv($A);                                                                                                                                                                                                                    
  1.321 μs (5 allocations: 6.19 KiB)                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    
                                                                                                                                                                                                                                          
julia> @btime inv($Asym);                                                                                                                                                                                                                 
  723.881 ns (5 allocations: 3.95 KiB)                                                                                                                                                                                                    

# on PR                                                                                                                                                                                                                                          
julia> @btime inv($A);                                                                                                                                                                                                                    
  564.087 ns (0 allocations: 0 bytes)                                                                                                                                                                                                     
                                                                                                                                                                                                                                          
julia> @btime inv($Asym);                                                                                                                                                                                                                 
  187.091 ns (0 allocations: 0 bytes)  
```